requirements.txt
fastapi
uvicorn[standard]
pyjwt
bcrypt
sqlalchemy
pyodbc
azure-storage-blob
azure-storage-queue
python-dotenv
